requirements.txt
aiosqlite==0.21.0
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.6.15
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pillow==11.2.1
pillow_heif==0.22.0
python-dotenv==1.1.1
python-telegram-bot==22.1
sniffio==1.3.1
typing_extensions==4.14.0
tzlocal==5.3.1
